Tutorial 1: Introduction to PyGMT for GeoScientists (#4)
Covers the basics of importing PyGMT into Python, and making a simple map (of New Zealand). Then we learn to plot a map of Oceania using the earth relief grids, and workshop users will have a chance to explore making a map of where they came from. Also covers saving the figure and jupyter notebook.